refactor: Calculator 패키지 리팩토링
파일 정렬 및 디렉토리 구조 리팩토링
