+++ src/main/java/calculator/Application.java
@@ -1,14 +1,14 @@
 package calculator;
 
-import calculator.service.Calculator;
+import calculator.domain.Calculator;
 import calculator.ui.ExpressionInput;
 import java.util.List;
 
 public class Application {
 
     public static void main(String[] args) {
-        List<String> formula = ExpressionInput.getExpression();
-        double result = Calculator.run(formula);
+        List<String> expression = ExpressionInput.getExpression();
+        double result = Calculator.run(expression);
         System.out.println(result);
     }
 }
