+++ src/main/java/calculator/domain/Operation.java
@@ -1,4 +1,4 @@
-package calculator.util;
+package calculator.domain;
 
 import java.util.Arrays;
